refactor: split the parameter migration ladder into per-area step modules
paramMigration.ts had 1290 code lines. migrateParams stays, and the wall,
lid, colour, cutout, text and compartment step functions move to sibling
modules, with the two shared guards in paramMigrationHelpers. Every step is
called in the same order as before and the module's exports are unchanged.
